adds distribution criteria
an example of distribution as a competitive advantage is Skype's button on Kazaa that allowed it to beat the hundreds of other VoIP startups in the market. In an enterprise sales company like Banksight, this could be deep customer relationships held by an experienced executive. This criteria was inspired by AlphaSOC's distribution through the Splunk app store, which cheaply generates leads to F100 companies.