Create Collabos - Reinvent Freelance
My apologies for contacting you this way, but I couldn't find an email and I've already sent a LinkedIn InMail to all contact people from Bloomberg Beta. Plus, you mentioned in your guide that we should be creative, so here it is.

We want to make sure you notice us as your mission to invest in companies that change how people work is close to ours. In fact, our platform is only dedicated to the process of work itself and nothing else.